Corrigido durante a apresentação

+++ src/manipulacaoDados/ManipulaDados.java
@@ -67,7 +67,7 @@ public class ManipulaDados {
 		String linhaArquivo = leitorArquivo.ler();
 		while (linhaArquivo != null) {
 			informacoes = linhaArquivo.split(";");
-			pessoa = new Pessoa(Float.parseFloat(informacoes[0]), informacoes[1], informacoes[2].charAt(0),
+			pessoa = new Pessoa(Long.parseLong(informacoes[0]), informacoes[1], informacoes[2].charAt(0),
 					Integer.parseInt(informacoes[3]), informacoes[4], informacoes[5], informacoes[6]);
 			populaListas(informacoes, pessoa);
 			populaArvore(pessoa);
@@ -143,7 +143,7 @@ public class ManipulaDados {
 	}
 	
 	public void pessoaNova(String[] dados) {
-		Pessoa pessoaNova = new Pessoa(Float.parseFloat(dados[0]), dados[1], dados[2].charAt(0), Integer.parseInt(dados[3]), dados[4],
+		Pessoa pessoaNova = new Pessoa(Long.parseLong(dados[0]), dados[1], dados[2].charAt(0), Integer.parseInt(dados[3]), dados[4],
 				dados[5], dados[6]);
 		populaListas(dados, pessoaNova);
 		populaArvore(pessoaNova);
